+++ TableProMobile/TableProWidget/QuickConnectWidget.swift
@@ -1,7 +1,6 @@
 import SwiftUI
 import WidgetKit
 
-@main
 struct QuickConnectWidget: Widget {
     let kind = "com.TablePro.QuickConnect"
 
@@ -13,5 +12,13 @@ struct QuickConnectWidget: Widget {
         .configurationDisplayName("Quick Connect")
         .description("Quickly connect to your databases.")
         .supportedFamilies([.systemSmall, .systemMedium])
+    }
+}
+
+@main
+struct TableProWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        QuickConnectWidget()
+        QueryLiveActivityWidget()
     }
 }